Fixed a pretty critical bug that caused spurious email to be sent to the list. Here is the entire text of an email I sent to Rasmus describing the bug in detail:
It seems that crappy logic in emailcron.py worked together with me pushing a bad version of make-index3.1.py caused the problem.

The way that it works is that make-index.py makes the emails and emailcron.py sends them. There is no state for make-index.py to know that the email has already been sent, so whenever it runs, it removes all emails and then remakes all future emails. emailcron is responsible for deciding which ones to send when. emailcron makes a file "mark_<the-date>" when it sends the file and won't resend the file if a file named "mark_<the-date>". It tries to send any email that is at most two days in the future. It also removes "mark"s in the past. make-index also runs on a scheduler to make sure that completed talks get moved to "past talks" promptly after they occur.

Do you see the horrible logic error? emailcron removes old "mark" files, but makes no effort to remove old emails. They only get removed by make-index.

So what happened here is that I pushed a broken version of make-index, and it failed to run so Eva's talk didn't get moved, and the email file didn't get deleted. Yesterday, when emailcron ran it noticed the mark and so didn't send the email, but it also deleted the mark because the associated date was in the past. Then today, there was no mark anymore and there was an email file dated at most two days into the future (actually in the past) that had no associated mark, so it sent it and marked it. It was going to continue sending the email every other day until I fixed make-index or fixed the bug.

The fix is that now emailcron also removes old emails, and it also checks whether the date is in the future before trying to send any emails.

It's actually sort of surprising that it took this long for this bug to come to light. Also, the problem with the bad version of make-index? It was a whitespace error >_<